1. move src/proxy/Json ---> include/Json    json 相关头文件不能放在源码路径下 2. 增加jsoncpp 对 long long 的支持

+++ PayGate/src/proxy/CHelper_pool.h
@@ -14,14 +14,14 @@ public:
 	CHelperPool();
 	~CHelperPool();
 public:
-	map<unsigned int, CHelperUnit*> 		m_helpermap;		//svid对应AllocServer
-	vector<unsigned int> 					m_svidlist;
-	map<unsigned int, CDecoderUnit*> 		m_objmap; 			/* client_unit */
-	vector<unsigned int> 					m_cmdlist;
-	map<unsigned int, std::vector<int> > 	m_levelmap;			//level对应svid
-	map<string, string> 					m_ipmap;
-	map<short, int> 						m_LevelCountMap;	//level---count 每个等级场的人数
-	vector<int> 							m_whitelist;        //容错命令白名单
+	map<int, CHelperUnit*> 				m_helpermap;		//svid对应AllocServer
+	vector<int> 						m_svidlist;
+	map<unsigned long, CDecoderUnit*> 	m_objmap; 			/* client_unit */
+	vector<int> 						m_cmdlist;
+	map<int, std::vector<int> > 		m_levelmap;			//level对应svid
+	map<string, string> 				m_ipmap;
+	map<short, int> 					m_LevelCountMap;	//level---count 每个等级场的人数
+	vector<int> 						m_whitelist;        //容错命令白名单
 };
 
 HTTP_SVR_NS_END
